refactor: :building_construction: pageディレクトリをcore配下に移動
