+++ lib/core/page/definition_post_page.dart
@@ -2,12 +2,12 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../core/router/app_router.dart';
-import '../feature/definition/application/definition_for_write_notifier.dart';
-import '../feature/definition/domain/definition_for_write.dart';
-import '../feature/definition/presentation/write_definition_base_page.dart';
-import '../feature/definition/util/after_post_navigation_type.dart';
-import '../util/mixin/presentation_mixin.dart';
+import '../../feature/definition/application/definition_for_write_notifier.dart';
+import '../../feature/definition/domain/definition_for_write.dart';
+import '../../feature/definition/presentation/write_definition_base_page.dart';
+import '../../feature/definition/util/after_post_navigation_type.dart';
+import '../../util/mixin/presentation_mixin.dart';
+import '../router/app_router.dart';
 
 /// 定義を投稿するページ
 @RoutePage()
