+++ lib/core/page/dictionary_sub_index_page.dart
@@ -2,11 +2,11 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../core/router/app_router.dart';
-import '../feature/user_profile/presentation/dictionary_author_widget.dart';
-import '../feature/word/presentation/index_tile.dart';
-import '../feature/word/util/dictionary_page_type.dart';
-import '../util/constant/initial_main_group.dart';
+import '../../feature/user_profile/presentation/dictionary_author_widget.dart';
+import '../../feature/word/presentation/index_tile.dart';
+import '../../feature/word/util/dictionary_page_type.dart';
+import '../../util/constant/initial_main_group.dart';
+import '../router/app_router.dart';
 
 @RoutePage()
 class DictionarySubIndexPage extends ConsumerWidget {
