+++ lib/core/page/setting_page.dart
@@ -3,15 +3,15 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../core/common_provider/in_app_review_provider.dart';
-import '../../../core/common_provider/launch_url.dart';
-import '../../../core/common_widget/shimmer_widget.dart';
-import '../../../core/router/app_router.dart';
-import '../../../util/constant/url.dart';
-import '../feature/auth/application/auth_state.dart';
-import '../feature/setting/presentation/delete_account_button.dart';
-import '../feature/user_config/application/user_config_state.dart';
-import '../feature/user_profile/application/user_profile_state.dart';
+import '../../../../core/common_provider/in_app_review_provider.dart';
+import '../../../../core/common_provider/launch_url.dart';
+import '../../../../core/common_widget/shimmer_widget.dart';
+import '../../../../core/router/app_router.dart';
+import '../../../../util/constant/url.dart';
+import '../../feature/auth/application/auth_state.dart';
+import '../../feature/setting/presentation/delete_account_button.dart';
+import '../../feature/user_config/application/user_config_state.dart';
+import '../../feature/user_profile/application/user_profile_state.dart';
 
 @RoutePage()
 class SettingRouterPage extends AutoRouter {
