+++ lib/core/page/word_list_page.dart
@@ -4,15 +4,15 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../core/common_widget/button/post_definition_fab.dart';
-import '../core/common_widget/infinity_scroll_widget.dart';
-import '../core/common_widget/simple_widget_for_empty.dart';
-import '../feature/word/domain/word.dart';
-import '../feature/word/presentation/word_tile.dart';
-import '../feature/word/presentation/word_tile_shimmer.dart';
-import '../feature/word_list/application/word_list_state_by_initial.dart';
-import '../util/constant/initial_main_group.dart';
-import '../util/extension/scroll_controller_extension.dart';
+import '../../feature/word/domain/word.dart';
+import '../../feature/word/presentation/word_tile.dart';
+import '../../feature/word/presentation/word_tile_shimmer.dart';
+import '../../feature/word_list/application/word_list_state_by_initial.dart';
+import '../../util/constant/initial_main_group.dart';
+import '../../util/extension/scroll_controller_extension.dart';
+import '../common_widget/button/post_definition_fab.dart';
+import '../common_widget/infinity_scroll_widget.dart';
+import '../common_widget/simple_widget_for_empty.dart';
 
 @RoutePage()
 class WordListPage extends ConsumerWidget {
