+++ lib/core/router/app_router.dart
@@ -9,28 +9,28 @@ import '../../feature/definition/presentation/write_definition_base_page.dart';
 import '../../feature/definition/util/after_post_navigation_type.dart';
 import '../../feature/definition_list/presentation/individual_dictionary_definition_list.dart';
 import '../../feature/word/util/dictionary_page_type.dart';
-import '../../page/base_page.dart';
-import '../../page/definition_detail_page.dart';
-import '../../page/definition_edit_page.dart';
-import '../../page/definition_post_page.dart';
-import '../../page/dictionary_everyone_page.dart';
-import '../../page/dictionary_individual_page.dart';
-import '../../page/dictionary_sub_index_page.dart';
-import '../../page/home_page.dart';
-import '../../page/introduction_page.dart';
-import '../../page/license_page.dart';
-import '../../page/profile_edit_page.dart';
-import '../../page/profile_top_page.dart';
-import '../../page/setting_page.dart';
-import '../../page/user_list_following_or_follower_page.dart';
-import '../../page/user_list_liked_page.dart';
-import '../../page/user_list_muted_page.dart';
-import '../../page/user_search_page.dart';
-import '../../page/user_search_result_page.dart';
-import '../../page/word_list_page.dart';
-import '../../page/word_search_result_page.dart';
-import '../../page/word_top_page.dart';
 import '../../util/constant/initial_main_group.dart';
+import '../page/base_page.dart';
+import '../page/definition_detail_page.dart';
+import '../page/definition_edit_page.dart';
+import '../page/definition_post_page.dart';
+import '../page/dictionary_everyone_page.dart';
+import '../page/dictionary_individual_page.dart';
+import '../page/dictionary_sub_index_page.dart';
+import '../page/home_page.dart';
+import '../page/introduction_page.dart';
+import '../page/license_page.dart';
+import '../page/profile_edit_page.dart';
+import '../page/profile_top_page.dart';
+import '../page/setting_page.dart';
+import '../page/user_list_following_or_follower_page.dart';
+import '../page/user_list_liked_page.dart';
+import '../page/user_list_muted_page.dart';
+import '../page/user_search_page.dart';
+import '../page/user_search_result_page.dart';
+import '../page/word_list_page.dart';
+import '../page/word_search_result_page.dart';
+import '../page/word_top_page.dart';
 import 'first_launch_guard.dart';
 
 part 'app_router.g.dart';
